Fixed merge errors/warnings

+++ Core/Inc/fixp.h
@@ -17,7 +17,7 @@
 
 #define FIX16_SHIFT 16
 #define FIX16_MULT(a, b) ((int32_t)(((int64_t)(a) * (int64_t)(b)) >> FIX16_SHIFT))
-#define FIX16_DIV(a, b)  ( ((a) << FIX16_SHIFT) / b )
+#define FIX16_DIV(a, b)  ( (int32_t)(((int64_t)(a) << FIX16_SHIFT) / (int64_t)(b)) ) // Still uses 64-bit intermediate.
 #define FIX16_POW(a, exp_int) ({ \
     int32_t _base = (a); \
     int32_t _result = 1 << FIX16_SHIFT; \
